TA-2919 updated tests
