+++ test/jest/client/components/pages/district-office-page/district-office-page.test.js
@@ -62,68 +62,4 @@ describe('District Office page', () => {
     const content = await waitForElement(() => getByTestId('office-error'))
     expect(content).toBeInTheDocument()
   })
-  it('contains an Events Component', async () => {
-    const initialState = undefined
-    const enhancer = applyMiddleware(thunk)
-    const store = createStore(reducers, initialState, enhancer)
-
-    const mockOfficeResponse = {
-      "leadership": {},
-      "location": [{
-          "id": 11803,
-          "type": "location",
-          "city": "Appleton",
-          "email": "score.foxcities@scorevolunteer.org",
-          "fax": null,
-          "geocode": {
-              "id": 19718,
-              "type": "geocode",
-              "latitude": "44.262804",
-              "longitude": "-88.409144"
-          },
-          "hoursOfOperation": null,
-          "name": "Fox Cities SCORE",
-          "phoneNumber": "920-831-4904",
-          "state": "WI",
-          "streetAddress": "125 N. Superior Street",
-          "zipCode": 54911
-      }],
-      "mediaContact": {},
-      "officeService": {},
-      "officeType": "SCORE Business Mentor",
-      "pims": {
-          "id": "107126",
-          "type": "pims",
-          "location": "282725"
-      },
-      "relatedDisaster": {},
-      "summary": {},
-      "website": {
-          "url": "https://foxcities.score.org",
-          "title": ""
-      },
-      "type": "office",
-      "title": "Fox Cities SCORE",
-      "id": 16008,
-      "updated": 1562852606,
-      "created": 1543957391,
-      "langCode": "en"
-    }
-    const fetchRestContentStub = jest.spyOn(fetchContentHelper, 'fetchRestContent')
-    fetchRestContentStub.mockImplementationOnce(() => Promise.resolve(mockOfficeResponse))
-
-    const fetchSiteContentStub = jest.spyOn(fetchContentHelper, 'fetchSiteContent')
-    fetchSiteContentStub.mockImplementationOnce(() => Promise.resolve(eventsTestData))
-
-    const { getByTestId } = render(
-      <Provider store={store}>
-        <DistrictOfficePage params={{ officeId: '16008' }} />
-      </Provider>
-    )
-    let content = await waitForElement(() => getByTestId('events'))
-    expect(content).toBeInTheDocument()
-
-    content = await waitForElement(() => getByTestId('events-button'))
-    expect(content).toBeInTheDocument()
-  })
 })
