+++ test/jest/client/components/templates/district-office/district-office.test.js
@@ -1,7 +1,9 @@
 import React from 'react'
-import { render, cleanup } from 'react-testing-library'
+import { render, cleanup, waitForElement } from 'react-testing-library'
 import DistrictOffice from 'templates/district-office/district-office.jsx'
 import 'jest-dom/extend-expect'
+import * as fetchContentHelper from 'client/fetch-content-helper.js'
+import eventsTestData from '../../test-data/events.json'
 
 afterEach(cleanup)
 
@@ -28,5 +30,59 @@ describe('District Office template', () => {
 
     const lenderMatch = getByTestId('office-lender-match')
     expect(lenderMatch).toBeInTheDocument()
+  })
+  it('contains an Events Component', async () => {
+    const mockOfficeResponse = {
+      "leadership": {},
+      "location": [{
+          "id": 11803,
+          "type": "location",
+          "city": "Appleton",
+          "email": "score.foxcities@scorevolunteer.org",
+          "fax": null,
+          "geocode": {
+              "id": 19718,
+              "type": "geocode",
+              "latitude": "44.262804",
+              "longitude": "-88.409144"
+          },
+          "hoursOfOperation": null,
+          "name": "Fox Cities SCORE",
+          "phoneNumber": "920-831-4904",
+          "state": "WI",
+          "streetAddress": "125 N. Superior Street",
+          "zipCode": 54911
+      }],
+      "mediaContact": {},
+      "officeService": {},
+      "officeType": "SCORE Business Mentor",
+      "pims": {
+          "id": "107126",
+          "type": "pims",
+          "location": "282725"
+      },
+      "relatedDisaster": {},
+      "summary": {},
+      "website": {
+          "url": "https://foxcities.score.org",
+          "title": ""
+      },
+      "type": "office",
+      "title": "Fox Cities SCORE",
+      "id": 16008,
+      "updated": 1562852606,
+      "created": 1543957391,
+      "langCode": "en"
+    }
+
+    const fetchSiteContentStub = jest.spyOn(fetchContentHelper, 'fetchSiteContent')
+    fetchSiteContentStub.mockImplementationOnce(() => Promise.resolve(eventsTestData))
+
+    const { getByTestId } = render(<DistrictOffice office={mockOfficeResponse} />)
+    let content = await waitForElement(() => getByTestId('events'))
+    expect(content).toBeInTheDocument()
+
+    content = await waitForElement(() => getByTestId('events-button'))
+    expect(content).toBeInTheDocument()
   })
 })
